refactor: clean up and restructure original story
